+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 ## [Unreleased]
 
+## [0.2.0] - 2025-09-13
+
+### Added
+- Added `have_empty_metadata` and `have_matching_metadata` matchers
+
+### Changed
+- Metadata matching must be done via the methods above instead of the fault matchers
+
 ## [0.1.0] - 2025-08-24
 
 - Initial release